refactor: 인수테스트 패키지 추가
